initialize options before calculation
